+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - ref(reactnative): Use clack prompts and share common code (dirty repo, login) (#473)
 - feat(reactnative): Add telemetry (#477)
+- feat(reactnative): Improve `build.gradle` patch so that it's more likely to work without changes in monorepos (#478)
 
 ## 3.15.0